move BetterNodeFinder to Utils package
