+++ packages/Sensio/src/Rector/FrameworkExtraBundle/TemplateAnnotationRector.php
@@ -12,11 +12,11 @@ use PhpParser\Node\Stmt\Return_;
 use Rector\BetterPhpDocParser\NodeAnalyzer\DocBlockAnalyzer;
 use Rector\Node\MethodCallNodeFactory;
 use Rector\Node\NodeFactory;
-use Rector\NodeTraverserQueue\BetterNodeFinder;
 use Rector\Rector\AbstractRector;
 use Rector\RectorDefinition\CodeSample;
 use Rector\RectorDefinition\RectorDefinition;
 use Rector\Sensio\Helper\TemplateGuesser;
+use Rector\Utils\BetterNodeFinder;
 
 final class TemplateAnnotationRector extends AbstractRector
 {
