+++ packages/Utils/composer.json
@@ -1,0 +1,25 @@
+{
+    "name": "rector/utils",
+    "description": "Util services like NodeFinder.",
+    "license": "MIT",
+    "authors": [
+        { "name": "Tomas Votruba", "email": "tomas.vot@gmail.com", "homepage": "https://tomasvotruba.com" }
+    ],
+    "require": {
+        "php": "^7.1",
+        "nikic/php-parser": "^4.0.3"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^7.2"
+    },
+    "autoload": {
+        "psr-4": {
+            "Rector\\Utils\\": "src"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Rector\\Utils\\Tests\\": "tests"
+        }
+    }
+}
