+++ packages/Utils/src/BetterNodeFinder.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\NodeTraverserQueue;
+namespace Rector\Utils;
 
 use PhpParser\Node;
 use PhpParser\NodeFinder;
